Reorganized browser detection and configuration logic (added UserAgent to verify what the platform can and cannot do).
Minor fixes for scripts (we were losing suspended scripts in one place but not other)
Scripts are suspended on world exit now (there is a config option to enable it). Firefox was
failing while manipulating SVG objects that were not in an SVG element (because its world was
not being displayed). Ultimately Firefox should be fixed :) and script suspension behavior should
be specifiable per action (?)

Major trimming of prototype.js